src/pyproject.toml
[tool.poetry]
name = "challenge-repo"
version = "0.1.0"
description = "challenge-repo"
authors = ["AI Ops <platform-aiops@zenjob.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "0.29.*"
fastapi = "0.111.*"
requests = "^2.32.3"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
langsmith = "^0.1.107"
langserve = ">=0.0.16"
langchain = "0.2.*"
langchain-core = "0.2.*"
langchain-community = "0.2.*"
langchain-openai = "^0.1.7"
langgraph = "^0.2.13"
sse-starlette = "^2.1.3"
#AWS
boto3 = "^1.34.11"

[tool.poetry.group.test.dependencies]
pytest-dotenv = "^0.5.2"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
parameterized = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

virtualenvs.in-project = true